+++ get_oslog.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
-log show --predicate 'subsystem == "com.meapri.hangyeol.inputmethod"' --last 15m > oslog.txt
+log show --predicate 'subsystem == "com.thlim.inputmethod.Hangyeol"' --last 15m > oslog.txt
 grep -v "Finder" oslog.txt | tail -n 30